requirements.txt
Django==1.9.5
djangorestframework==3.3.3
requests==2.9.1
selenium==2.53.1
